Re-organized packages; Clean-up preferences code.
